requirements.txt
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1